requirements.txt
requests
requests_cache
beautifulsoup4
prettytable
pycryptodome